Auditing Data

+++ backend/src/main/java/com/gabriel/dscatalog/services/CategoryService.java
@@ -3,8 +3,11 @@ package com.gabriel.dscatalog.services;
 import com.gabriel.dscatalog.dto.CategoryDTO;
 import com.gabriel.dscatalog.models.Category;
 import com.gabriel.dscatalog.repositories.CategoryRepository;
+import com.gabriel.dscatalog.services.exceptions.DatabaseException;
 import com.gabriel.dscatalog.services.exceptions.EntityNotFound;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -43,17 +46,19 @@ public class CategoryService {
     @Transactional
     public CategoryDTO update(Long id,CategoryDTO dto) {
             Optional<Category> obj = repository.findById(id);
-            Category entity = obj.orElseThrow(() -> new EntityNotFound("Entity not found"));
+            Category entity = obj.orElseThrow(() -> new EntityNotFound("Id not found" + id));
             entity.setName(dto.getName());
             entity = repository.save(entity);
             return new CategoryDTO(entity);
     }
 
     public void delete(Long id) {
-        try {
+        if (!repository.existsById(id)){
+            throw new EntityNotFound("Id not found " + id);
+        }try{
             repository.deleteById(id);
-        } catch (EntityNotFound e) {
-            throw new EntityNotFound("Entity not found");
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Integrity violation");
         }
     }
 }
